requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
Pillow==7.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.12.0
requests==2.23.0
urllib3==1.25.9